pyproject.toml
[project]
name = "docx-link-renderer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "python-docx>=1.1.2",
    "requests>=2.32.4",
]
